requirements.txt
numpy~=1.19.1
pillow~=7.2.0
scikit-image~=0.16.2